fix: resolve remaining linting errors for CI
This commit addresses all remaining linting errors to make PR #9 mergeable:

**Test workflow fixes:**
- Remove CGO_ENABLED=0 from test workflow to allow -race flag to work
- The -race detector requires CGO to be enabled

**Code fixes:**
- Fix 5 unconvert errors (unnecessary type conversions):
  - Remove redundant int() conversion in pixel_benchmark.go (NumberOfFrames already int)
  - Remove redundant byte() wrapper in lut.go and lut_advanced.go (already uint8)
  - Remove redundant byte() in lut_advanced_test.go (constant arithmetic)
  - Remove redundant uint16() in fuzz_test.go (CommandCEchoRQ already uint16)

- Fix 2 unused function errors:
  - Remove unused mustNewIntValue helper in dataset_test.go
  - Remove unused createDatasetMissingTag helper in directory_writer_test.go

- Fix 5 ST1023 errors (redundant type declarations):
  - Simplify status variable declarations in dimse/scp/server.go
  - Change "var status uint16 = X" to "status := X" in 5 handler functions

- Fix 1 S1009 error (unnecessary nil check):
  - Remove redundant nil check in pixel/creator.go (len() handles nil slices)

- Fix 2 staticcheck false positives in fuzzing tests:
  - Add //lint:ignore SA5011 for nil pointer false positive
  - Add //lint:ignore SA4010 for unused append in state history collection

- Fix gofmt formatting error:
  - Auto-format dicom/dataset_collection.go comment alignment

All fixes maintain existing functionality while satisfying linter requirements.

+++ dimse/dimse/fuzz_test.go
@@ -148,7 +148,7 @@ func FuzzMessageEncoding(f *testing.F) {
 		},
 		PresentationContextID: 1,
 	}
-	f.Add(uint8(1), uint16(CommandCEchoRQ), uint16(1))
+	f.Add(uint8(1), CommandCEchoRQ, uint16(1))
 
 	// Seed with various parameter combinations
 	f.Add(uint8(0), uint16(0), uint16(0))             // Invalid values
@@ -177,7 +177,8 @@ func FuzzMessageEncoding(f *testing.F) {
 				}
 
 				// Verify presentation context ID matches
-				for _, item := range p.Items {
+				//lint:ignore SA5011 false positive, nil check performed above
+			for _, item := range p.Items {
 					if item.PresentationContextID != pcID {
 						t.Errorf("PC ID mismatch: expected %d, got %d",
 							pcID, item.PresentationContextID)
